ClusterTasks for cosign verify, verify-attestation
This adds two ClusterTasks to run `cosign verify` and
`cosign verify-attestation` as Tekton tasks, i.e. within the cluster.

The Kaniko demo script is changed to illustrate running these tasks if
the image is pushed to the internal OpenShift registry, this is just a
shortcut, we could use a parameter or run both on the end user computer
and on the cluster.

Further work should include rebasing this on
redhat-appstudio/build-definitions#28 and figure out if these tasks
should be part of `components/build` tasks.

Ref. https://issues.redhat.com/browse/HACBS-106